Discontinued use of fullSorter.py after text comparisons with sorter.py equivalents.
Seems like there are some errors with the way that fullSorter.py
outputs files. This is most likely an issue having to do with the
actual laws site itself and not the code, as the chapter-by-chapter
versions seem to not have this issue yet use the same algorithms to
process data.